+++ src/qcommon/q_unicode.h
@@ -68,8 +68,8 @@ void Q_UTF8_FreeFont(fontHelper_t *font);
 char *Q_UTF8_Encode(unsigned long codepoint);
 int Q_UTF8_Store(const char *s);
 char *Q_UTF8_Unstore(int e);
-void Q_UTF8_ToUTF32(char *string, int *charArray, int *outLen);
-void Q_UTF32_ToUTF8(uint32_t *charArray, char *string, int *outLen);
+void Q_UTF8_ToUTF32(char *string, uint32_t *charArray, size_t *outLen);
+void Q_UTF32_ToUTF8(uint32_t *charArray, char *string, size_t *outLen);
 
 size_t Q_EscapeUnicode(char *fromStr, char *toStr, size_t maxSize);
 size_t Q_UnescapeUnicode(char *fromStr, char *toStr, size_t maxSize);
